% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{failwith}
\alias{failwith}
\title{Function operator to catch exceptions for optimisation when fitting neural network.
Courtesy Hadley Wickam.}
\usage{
failwith(default = NULL, f, quiet = TRUE)
}
\arguments{
\item{default}{default value to return when exception occurs}

\item{f}{function to try}

\item{quiet}{logical. Should the functin fail silently?}
}
\description{
Function operator to catch exceptions for optimisation when fitting neural network.
Courtesy Hadley Wickam.
}

