% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forward-propogate.R
\name{forwardProp}
\alias{forwardProp}
\title{forwardProp performs a forward propogation through the network,
using current value of parameters Theta}
\usage{
forwardProp(unrollThetas, Thetas, nUnits, nLayers, lambda, outcome, a, z,
  gradient, delta, Deltas)
}
\arguments{
\item{unrollThetas}{vector comprised of unrolled parameter matrix elements}

\item{Thetas}{template list of matrices allocated to correct size}

\item{nUnits}{number of units in hidden layers}

\item{nLayers}{number of hidden layers}

\item{lambda}{penalty ("weight-decay") term}

\item{outcome}{matrix of 'dummied' outcomes}

\item{a}{template list of activation matrices (with "zeroth" layer filled in with inputs)}

\item{z}{template list of transformed activation matrices}

\item{gradient}{template list of matrices of gradients}

\item{delta}{template list of matrices of errors}

\item{Deltas}{template list of matrices of product of activations and errors}
}
\description{
forwardProp performs a forward propogation through the network,
using current value of parameters Theta
}

