% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forward-propogate.R
\name{forwardPropogate}
\alias{forwardPropogate}
\alias{forwardPropogate_c}
\title{forwardPropogate performs a forward propogation through the network,
using current value of parameters Theta}
\usage{
forwardPropogate(Thetas, a, nLayers)

forwardPropogate_c(Thetas, a, nLayers)
}
\arguments{
\item{Thetas}{list of matrices of parameters}

\item{a}{template of activation matrices (with "zeroth" layer filled in with inputs)}

\item{nLayers}{number of hidden layers}
}
\description{
forwardPropogate performs a forward propogation through the network,
using current value of parameters Theta
}
\section{Functions}{
\itemize{
\item \code{forwardPropogate_c}: 
}}

