% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\docType{methods}
\name{getParams}
\alias{getParams}
\alias{getParams,nnePtR-method}
\title{Set the generic for accessor (getter) for fitted coefficients}
\usage{
getParams(object)

\S4method{getParams}{nnePtR}(object)
}
\arguments{
\item{object}{object of class nnePtR}
}
\description{
Set the generic for accessor (getter) for fitted coefficients
}
\section{Methods (by class)}{
\itemize{
\item \code{nnePtR}: 
}}

