% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-definition.R
\docType{methods}
\name{initialize,nnePtR-method}
\alias{initialize,nnePtR-method}
\title{Initializor to catch input errors}
\usage{
\S4method{initialize}{nnePtR}(.Object, input, outcome, levels, n_layers,
  n_units, penalty, cost, fitted_params, info)
}
\arguments{
\item{.Object}{object of class nnePtR}

\item{input}{matrix of inputs. Dimensions are (n training samples x n features)}

\item{outcome}{factor variable for outcome}

\item{levels}{character vector of levels of outcome}

\item{n_layers}{number of hidden layers in the network}

\item{n_units}{number of units in the hidden layers}

\item{penalty}{penalty term (weight decay)}

\item{cost}{final cost obtained through optimisation}

\item{fitted_params}{list of matrices holding the fitted parameters}

\item{info}{list containing infomation from optimisation}
}
\description{
Initializor to catch input errors
}

