% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-definition.R
\docType{class}
\name{nnePtR-class}
\alias{nnePtR-class}
\title{S4 class definition for a neural network}
\description{
S4 class definition for a neural network
}
\section{Slots}{

\describe{
\item{\code{train_input}}{}

\item{\code{train_outcome}}{}

\item{\code{levels}}{}

\item{\code{n_layers}}{}

\item{\code{n_units}}{}

\item{\code{penalty}}{}

\item{\code{cost}}{}

\item{\code{lambda}}{}

\item{\code{fitted_params}}{}

\item{\code{info}}{}
}}

