% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-definition.R
\name{nnetBuild}
\alias{nnetBuild}
\title{Constructor for nnePtR}
\usage{
nnetBuild(train_input, train_outcome, nLayers = 1, nUnits = 25,
  lambda = 0.01, seed = 1234, iters = 100, optim_method = "L-BFGS-B")
}
\arguments{
\item{train_input}{data frame or matrix of input features}

\item{train_outcome}{vector of outcome. should be a factor}

\item{nLayers}{number of hidden layers}

\item{nUnits}{number of units in hidden layers. Constant across each hidden layer.}

\item{lambda}{penalty term}

\item{seed}{seed for initialising network}

\item{iters}{number of iterations (passed to optim)}

\item{optim_method}{opitimisation method (passed to optim)}
}
\description{
Constructor for nnePtR
}

