% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{predict,nnePtR-method}
\alias{predict,nnePtR-method}
\title{Predict generic}
\usage{
\S4method{predict}{nnePtR}(object, newdata, type = "response")
}
\arguments{
\item{object}{object of class nnePtR}

\item{newdata}{data to generate predictions for}

\item{type}{select type = "response" for predicted class, or "prob" for probabilities}
}
\description{
Predict generic
}

