% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forward-propogate.R
\name{propogate}
\alias{propogate}
\alias{propogate_c}
\title{propogate performs the meat of the work for forwardProp}
\usage{
propogate(Thetas, a, nUnits, nLayers)

propogate_c(Thetas, a, nUnits, nLayers)
}
\arguments{
\item{Thetas}{list of matrices of parameters}

\item{a}{template of activation matrices (with "zeroth" layer filled in with inputs)}

\item{nUnits}{number of units in hidden layers}

\item{nLayers}{number of hidden layers}
}
\description{
propogate performs the meat of the work for forwardProp
}
\section{Functions}{
\itemize{
\item \code{propogate_c}: 
}}

