% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{rollParams}
\alias{rollParams}
\title{Roll a vector of parameters into a list of matrices
defined by a template}
\usage{
rollParams(x, nLayers, Thetas_size)
}
\arguments{
\item{x}{input vector to be rolled into list of matrices}

\item{nLayers}{number of hidden layers in neural network}

\item{Thetas_size}{template list of matrices for x to be rolled into}
}
\description{
Roll a vector of parameters into a list of matrices
defined by a template
}

