% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{sigmoid}
\alias{sigmoid}
\alias{sigmoid_c}
\title{Sigmoid function}
\usage{
sigmoid(x)

sigmoid_c(x)
}
\arguments{
\item{x}{input}
}
\description{
Sigmoid function
}
\section{Functions}{
\itemize{
\item \code{sigmoid_c}: 
}}

