% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{sigmoidGradient}
\alias{sigmoidGradient}
\alias{sigmoidGradient_c}
\title{Sigmoid gradient function}
\usage{
sigmoidGradient(x)

sigmoidGradient_c(x)
}
\arguments{
\item{x}{input}
}
\description{
Sigmoid gradient function
}
\section{Functions}{
\itemize{
\item \code{sigmoidGradient_c}: 
}}

