% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{splitfn}
\alias{splitfn}
\title{Split function. As gradient and cost both calculated in backProp function
we want to cache values as optim will call the functions seperately (for fn and gr).
This closure splits the function, and will check if x (the par argument of optim)
changes between calls. Will retrieve cached values if not (i.e will calculate cost
and gradient for for fn, and retrieve cached values for gr)}
\usage{
splitfn(f)
}
\arguments{
\item{f}{the function to be split}
}
\description{
Split function. As gradient and cost both calculated in backProp function
we want to cache values as optim will call the functions seperately (for fn and gr).
This closure splits the function, and will check if x (the par argument of optim)
changes between calls. Will retrieve cached values if not (i.e will calculate cost
and gradient for for fn, and retrieve cached values for gr)
}

