% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{summary,nnePtR-method}
\alias{summary,nnePtR-method}
\title{Show generic}
\usage{
\S4method{summary}{nnePtR}(object)
}
\arguments{
\item{object}{object of class nnePtR}
}
\description{
Show generic
}

