% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-definition.R
\docType{methods}
\name{initialize,nnePtR-method}
\alias{initialize,nnePtR-method}
\title{Initializor to catch input errors}
\usage{
\S4method{initialize}{nnePtR}(.Object, input, outcome, n_layers, n_units,
  penalty, cost, fitted_params, info)
}
\arguments{
\item{.Object}{object of class nnePtR}

\item{n_layers}{}

\item{n_units}{}

\item{penalty}{}

\item{cost}{}

\item{fitted_params}{}

\item{info}{}

\item{train_input}{}

\item{train_outcome}{}

\item{lambda}{}
}
\description{
Initializor to catch input errors
}

