% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{splitfn}
\alias{splitfn}
\title{split function - make a closure to speed up use of optim by caching. now we just need backprop!
modify backprop so it calculates the cost.
Then first pass of backprop- calcs cost and gradient + caches. second call, retrieves cached gradient}
\usage{
splitfn(f)
}
\arguments{
\item{f}{f the function to be split}
}
\description{
split function - make a closure to speed up use of optim by caching. now we just need backprop!
modify backprop so it calculates the cost.
Then first pass of backprop- calcs cost and gradient + caches. second call, retrieves cached gradient
}

