% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{unrollParams}
\alias{unrollParams}
\alias{unrollParams_c}
\title{Unroll a list of matrices into a single vector}
\usage{
unrollParams(x)

unrollParams_c(x)
}
\arguments{
\item{x}{input list of matrices}
}
\description{
Unroll a list of matrices into a single vector
}
\section{Functions}{
\itemize{
\item \code{unrollParams_c}: 
}}

